Allow apps to override the application progress check
In addition to their Helm release readiness, applications sometimes
need to run custom operations to check if they have been fully
applied and are ready to be used. A lifecycle was implemented to
allow running customizable code when the application framework
checks whether the application operation was completed.

Test Plan:
PASS: Build app metrics-server with the call to the new
      lifecycle added.
PASS: Apply metrics-server and check if lifecycle is called before
      the apply completed operation.

Story: 2011242
Task: 51206

Change-Id: I5423bc5793cacc6cb41c44ae21d0c52f208fba5d
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>