Portieris webhook to ignore failures in upgrade
The webhook from portieris is configured to fail if the app cannot
respond, as a safety measure. However, during upgrades, this causes
issues when portieris itself and the other apps are updated.

This change modifies the webhook to ignore failures. The webhook will
be recreated during portieris upgrade/apply, going back to the
settings created using overrides.

Included backup of the webhook in the old version config folder. This
backup will be used in case of a rollback, and will be deleted upon
the end of the upgrade.

Also, cert-manager webhook sometimes fails after the app is updated,
during the period where the server is being reconfigurated. This is
a documented behavior, and it's recomended to verify that the creation
of certificates is working as a sanity check. Added this in the script
that updates the k8s apps.

This review also remove a unused script: 73-upgrade-portieris.sh

Test plan:
PASS: Performed stx 8 -> stx 10 upgrade activation. Observed that after
      the execution of the script 06, the portieris webhook is changed
      to ignore on failures.
      Observed that after portieris is applied, the webhook is recreated
      with the user settings.

PASS: Performed stx 8 -> stx 10 upgrade activation and stopped after
      the execution of the script. Triggered script with rollback
      action manually, and observed that the webhook is recreated
      using the backup.

Story: 2010676
Task: 51214

Change-Id: Ib4ddc91ed265bf2e4bee8b739697322ef9f77cd0
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>