Handle BlockingIOError on IPsec socket communication
It was experienced cases where a BlockingIOError is raised during the
receiving of messages on IPsec server per the execution of socket.recv
function. During this error, IPsec client has a prolonged execution on
response IPsec server and communication between IPsec server-client is
terminated.

To handle this exception, this commit adds logging and a wait time when
this error is experienced, in order to maintain IPsec communication and
IPsec be configured properly on nodes during server-client interaction.

Test Plan:
PASS: Deploy a standard system and upgrade from stx 9 to stx 10.
      Observe that IPsec is configured and enabled on all hosts, SAs
      are established and nodes are reachable.
PASS: Deploy a standard system and upgrade abort and rollback from
      release stx 10 to stx 9.
      Observe that IPsec is disabled on all hosts, cert, keys and
      conf files are not present on nodes during activate-rollback.
      All nodes are reachable and stable.
PASS: On a standard sytem, execute disable-ipsec.yml playbook and
      observe that IPsec is disabled on all nodes.
PASS: On a standard system, execute enable-ipsec.yml playbook and
      observe that IPsec is enabled and configured on all nodes.
PASS: On a standard system, execute enable-ipsec.yml and
      disable-ipsec.yml playbooks in order to enable, disable and
      re-enable IPsec. Observe that IPsec is re-enabled successfully
      and all nodes are configured.

Story: 2010940
Task: 51226

Change-Id: Idfac2cbed782e363c6e95ea1bec85aec194691ea
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>