Add rook-ceph recovery check in "restore-complete"
During the backup and restore process with rook-ceph configured,
depending on how the OSDs and monitors were distributed among the
hosts, the ceph-recovery jobs may only be started after the hosts
have been unlocked. Therefore, to prevent being able to perform a
"system restore-complete" command while it is still in the recovery
process, a check has been added to the conductor manager.

Test Plan:
- PASS: Configure Standard with OSDs on controller-1
        and worker-0, and monitors on the other hosts.
- PASS: Perform B&R
- PASS: After restore, unlock controller-1 only after
	    all hosts have been unlocked.
- PASS: Try running "system restore-complete" as soon
        as it is unlocked

Partial-Bug: 2086473

Change-Id: Iea70106c834cfa5c5a9407bfdc8f00194c86a1b9
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>