Improve check for duplicate subnet during route creation
When the system has a lot of routes, like when a large number of
subclouds are deployed, further route creations are slowed down due
to routes_get_by_host, which is a call that happens sequentially,
taking around 1.5 seconds to complete. This commit creates a new
optimized call to get only necessary information for the duplicate
subnet check, which reduces the call for around 100 ms.

Test plan:
  - PASS: Add 250 subclouds in parallel with already a large number of
          subclouds in db and verify there's no timeout during
          creation.
  - PASS: Deploy a SX. Add a route and verify it was created without
          issues.
  - PASS: Create a route with a duplicate subnet and verify it's
          correctly blocked.

Story: 2011106
Task: 51251

Change-Id: If7d5d79323dd63d392b75b7e12d121a10f8f6075
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>