Update kustomization file using atomic operation
Currently, the process of updating Kustomization files is not atomic,
which can lead to missing or empty files when sysinv is killed or the
machine is forcibly powered off. First, the old file is removed, then
another one is written and saved to the original path. This commit
changes that logic, allowing the updated file to be written first. Once
it is written to disk, the original file is replaced via os.rename,
which is an atomic operation that complies with POSIX requirements.

This commit also unifies the logic that reads Kustomization files
into a single method. In addition, an unnecessary call to os.remove was
removed when handling Kustomization files during the charts retrieval
process.

Test Plan:
PASS: AIO-SX fresh install.
PASS: Upgrade from previous release.
PASS: upload/apply/remove/delete/update metrics-server app
PASS: upload/apply/remove/delete/update dell-storage app

Closes-bug: 2087558

Change-Id: I94ce2d9030f7543722cf08061b9fca878e71e3bc
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>