Fix for alarm 500.200 triggered when reapplying KubeVirt App
While removing the KubeVirt app, the watch thread in the background
is triggered by the event change in the CDI resource. This causes
the watch_event to be set to True. As a result, when the KubeVirt
app is applied again, the event cannot be set a second time, leading
to a failure.

This update addresses the issue where Alarm 500.200 is
triggered when the KubeVirt app is removed and re-applied.

Test Plan:
PASS: Upload and apply the kubevirt-app, then verify that the
500.200 alarm for cdi-uploadserver-client-cert does not appear
after running the health status check.
PASS: Remove and re-apply the kubevirt-app, then verify that the
500.200 alarm for cdi-uploadserver-client-cert does not appear

Partial-Bug: 2084142

Change-Id: I0377e038461957c8161d7c0c0ded13862c2085ce
Signed-off-by: Boovan Rajendran <boovan.rajendran@windriver.com>