Change logic when deleting subcloud endpoints
After keystone endpoints are created in a subcloud during bootstrap,
if the subcloud is deleted and re-added, new endpoints will be created
for the new region, but old endpoints are not deleted. This commit
fixes this by changing the logic on endpoint deletion, deleting all
endpoints that do not belong to the current region.

Test plan:
  - PASS: Force a failure in subcloud bootstrap after the endpoints
          were created. Delete the subcloud and re-add. Verify new
          endpoints were created and old endpoints were deleted.
  - PASS: Finish the whole deployment and manage the subcloud. Verify
          it became in-sync after.

Closes-bug: 2087073

Change-Id: I1e48539ef3c8a9aee3a036c40d2a3754271502ee
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>