Allow subnet sharing for subclouds using the Admin network.
In the case of the admin network, it should be possible to 'share'
the subnet with multiple subclouds, as long as their admin address
pools do not overlap.

Currently, the restriction is to not allow subnets to be shared
at all. Since the admin subnet should only have a maximum of
three addresses (considering the AIO-DX case - floating, two
unit IP), for the subcloud controller(s), it does not make sense
to reserve the entire subnet range.

With these modifications, care must be taken to only delete the route
from the system controller to the subcloud subnet when all subclouds
using that subnet have been deleted.

Test Plan:

The following tests have been performed for a system with IPv4, IPv6,
and dual-stack enabled admin network.

- AIO-SX: bootstrap including only the admin_floating_address being
  specified.
- AIO-SX: bootstrap including only the admin_start_address and
  admin_end_address (being equal) being specified.
- AIO-DX: bootstrap including only the admin_floating_address being
  specified.
- AIO-DX: bootstrap including only the admin_start_address and
  admin_end_address being specified (default behaviour).
- AIO-SX and AIO-DX admin_subnet prefix len /30:
  - Ensure only AIO-SX pass.
- Install 3 AIO-SX subclouds, using different IP addresses on
  the same admin_subnet. Ensure:
  - A subcloud can be deleted. The route between system controller
    and the subclouds is not deleted until the last subcloud using
    the subnet is deleted.
  - A subcloud using the same admin address as the other subclouds
    cannot be added.
  - A subcloud can be deleted and re-added using the same IP address
    as it used before.
- Subcloud update. Change network from mgmt to admin
- Subcloud migration
  - to the same system controller.
  - to a different system controller.
- Subcloud remote install
- Subcloud factory install enrollment
- Subcloud AIO-SX to AIO-DX migration
- Subcloud AIO-SX and AIO-DX admin_subnet prefix len /30:
  - Ensure only AIO-SX pass.
- 2 Subclouds with one admin_subnet contained in the other
  - Ensure routes are created and deleted correctly
- AIO-SX, AIO-DX and subcloud upgrade/rollback single and dual-stack

Story: 2011191
Task: 51370

Change-Id: I889a11ff5d1e71d65bff3b3f4cc1b986f94199b4
Signed-off-by: Steven Webster <steven.webster@windriver.com>
Signed-off-by: Caio Bruchert <caio.bruchert@windriver.com>