Not enable HTTPS using cert-mon
To be able to re introduce switching the Openstack and webserver
endpoints to HTTP via CLI, this review modifies cert-mon to not enable
HTTPS when the Rest API / GUI (ssl) certificate is created/changed.

This way, the certificate will still be renewed by cert-manager and
installed as the current 'ssl' certificate, allowing the user to
switch from HTTPS to HTTP and from HTTP back to HTTPS at any time,
but not automatically switch from HTTP to HTTPS if the certificate
changes.

This review also allowed the system to change from HTTPS to HTTP
while the cert is managed by cert-manager, as cert-manager renewals
will not interfere w/ the system's HTTPS enabled status.

Test plan:
PASS: Bootstrap AIO-SX

PASS: Switch to 'https_enabled: false'
      Force renewal of REST API/GUI cert
      Observe that the system remains w/ HTTPS disabled

Story: 2011266
Task: 51332

Change-Id: I1940e471c9ea92a7eaa7175122bf8f048563031a
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>