Update misleading details message
This commit updates a misleading error details message that
displays that an upgrade is running when the message should
show that a software deployment is in progress.

Test Plan:

PASS: Deploy a software update in the system controller and power
off the subcloud. Attempt to delete the subcloud and verify that
the message displayed is "A software deployment is in progress
with state start."

Closes-bug: 2088086

Change-Id: I43d5ec2b0f2f0f94708881297970d31899a7e971
Signed-off-by: Gustavo Pereira <gustavo.lyrapereira@windriver.com>