Avoid recreating N3000 entries after reset
When N3000 is reset, the onboard devices (0d58) can have their PCI
address changed. With the fix, the database is updated when the PCI
address changes after an N3000 reset, rather than recreating the entry
for the same MAC address. This is done to avoid issues with the
Deployment Manager, as it uses the host-port UUID field as a reference
for configuration.

Test Plan
=========

[PASS] New unit tests using simplex and non-simplex simulation
[PASS] Configure pci-sriov interfaces on onboard N3000 devices
[PASS] Fresh installation with AIO-DX. Check if the host-port entries
are updated instead of recreated after unlocking.
[PASS] Controllers reboot
[PASS] Host lock/unlock

Closes-Bug: 2089264

Change-Id: Ib5fe7344d5cda9f36e61a43daa1644911b220524
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>