Fix for failed K8s upgrade on multiple subclouds.
The manifest apply fails on the first attempt because of the timeout
and it retries.
After retrying and successfully re-applying, the cp_version
(Control Plane Version) is not updated to the target version in time
The Kubernetes upgrade is aborted due to the cp_version mismatch.

This commit adds the code to retry 'kube_get_control_plane_versions'
10 times with the delay of 1 sec between each retry.
This retrying will only be done if the control plane image version
is not equal to the target image version

Test plan:
PASS: Deploy the latest ISO in both SX and DX environment.
PASS: Perform manual Kubernetes components upgrade from v1.28.4
      to v1.29.2.
PASS: Perform Kubernetes version upgrade using cloud orchestration
      strategy from v1.28.4 to v1.29.2.

Closes-bug: 2092310

Change-Id: Ifb03142b8262c4797ae80566b7c0ef04e3e324bf
Signed-off-by: rakshith mr <rakshith.mr@windriver.com>