Fix subclouds/external_oam_subnet_ip_family db field null after upgrade
The 100-update_dcmanager_subclouds_table.py upgrade script adds the
external_oam_subnet_ip_family field to the subclouds table and should
fill it afterwards, but the WHERE clause in the update query filters
for empty string and not null value, and none of the records gets
updated. This change fixes the query so that all records are correctly
updated.

Test plan

[PASS] Upgrade System Controller, check that table has correct values
[PASS] Update subcloud management network addresses

Closes-Bug: 2091661
Change-Id: Icbc83740e50c41a11af26f41cd8aa4d9a14b7c8f
Signed-off-by: Lucas Ratusznei Fonseca <lucas.ratuszneifonseca@windriver.com>