Fix for deletion and reinstall operation
This change is necessary to add a new state to the stor to avoid loss of
OSD information before deleting the OSD from the ceph cluster by the
Rook-Ceph application when the OSD is in configured state.

The Rook-Ceph application lifecycle will be responsible for deleting
the stor from the database and the ceph cluster, and also for
removing OSD function from host-fs functions. The new delete-with-app
state will require an application to be applied again in order
to implement the changes.

Now, there is no need to lock the host to delete the stor anymore.
Additionally, during the host reinstall process,
the state "configuring-with-app" is assigned to the storage being
reinstalled, so that it will be properly deleted and reconfigured
with the app apply in the future.

When the storage-backend is modified, the state will transition to
configuring-with-app to ensure that the ceph-rook backend does not
remain in the configured state. This is necessary because the
application must be reapplied to get the new changes.

Test Plan:
 - PASS: Check if the state of storage-backend ceph-rook-store is being
         updated to configuring-with-app
 - PASS: Execute host-stor-delete command and check if the stor state
         has been updated to deleting-with-app when it does not affect
         replication factor and is not with configuring-with-app state
 - PASS: Check if host-stor-delete command can be executed with host
         unlocked
 - PASS: Verify that the unit test has been updated to support the
         new behavior
 - PASS: Check if stors can not be deleted if the app is any progress
         status

Partial-Bug: 2093897

Change-Id: I530b7185b64c93a7d4292b7b3edfceb97140cf28
Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>