Move upgrade scripts to /usr/local/share
In the current directory /etc/upgrade.d, when upgrade scripts
are removed or renamed in a patch, the patching system using
apt-ostree does not recognize these changes because the files
are considered config files, and ends up with duplicated files
in the directory after the patch is applied.

This commit moves the upgrade scripts from /etc to /usr/local/share
so that scripts are changed correctly in the scenario described
previously.

TODO (another commit): change the patch scripts directory from
                       /etc/update.d to /usr/local/share/

Test Plan
PASS: AIO-SX: apply a patch with deleted, changed and renamed scripts
      in stx-10, verify the changes are applied accordingly
PASS: AIO-SX: build a prepatched iso with the changes from the
      previous TC and run a major release deployment stx-8 -> stx-10

Closes-bug: 2091944

Change-Id: Ia36dcda9630a9b3d7e37fe0ce82e3e1fcb09daa1
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>