Block lock/swact in deployment transient states
There are some states in USM deployment that are transient,
i.e. they run scripts that may take some time to complete, and
while in these states the user should be blocked from swacting
or locking the controllers.

This commit adds this check to the upgrade lock/swact semantic
check.

Test Plan
PASS: AIO-DX: attempt lock and swact with deployment in the
      transient states, verify the swact is blocked
PASS: AIO-DX: attempt lock and swact with deployment in other
      states, verify it is allowed
PASS: AIO-DX: upgrade using orchestration and verify the previous
      TCs above

Closes-bug: 2092237

Change-Id: I8b73eab668e0640a924142fa3524e21a251304d2
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>