Plugin from previous version of the app activated incorrectly.
During the upgrade, more precisely during the execution of script
21 in the activate process, it was observed that the old plugins
folder was recreated during the app update. This happened because
the evaluate_apps_reapply function was executed for that app during
its update.

If the app is evaluated for reapplication during the update, the
old plugin folder is recreated, causing any operations performed
with the apps to fail.

These changes fix this problem. If the app is updating, a skip in
the "for" is made inside evaluate_apps_reapply function so that
the problem does not occur again.

Test Plan:
PASS: build-pkgs && build-image
PASS: upgrade from stx-9 to master
PASS: 21-k8s-app-upgrade.sh script runs without errors

Closes-bug: 2092959

Change-Id: Idb5f1bfe9b22eb45d63e95e176b02428d4d8b7df
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>