Add delete armada deployment function to script 01
After some tests, was possible to see that
sometimes the script executions fails after
successfully deleting the helm release, while the
armada deployment remains in the system, as a
result, every attempt to delete the pods fails as
k8s raises another pod after the deletion.
This commit added to the armada script a function
to remove the deployment before the attempts to
delete the armada pods.

Test Plan:
PASS: AIO-DX stx-8 install
PASS: upgrade stx-8 to master and observe if script
      01 was executed with success during activate
      process.
PASS: Edit script 01 making the skipping the helm
      attempts to remove the resource fails and
      check if the script successfully remove the
      armada deployment.

Closes-Bug: 2093282
Change-Id: I7c37e6b4db6b50b9d0be150837b24fe4cb06fc99
Signed-off-by: Edson Dias <edson.dias@windriver.com>