Fix PV creation with null values in database
When deploying a new version and unlocking the standby host, the
device_node may be re-enumerated and change from /dev/sda to /dev/sdb
and vice-versa, for example.

Which causes the information in the DB about the partitions to diverge
as the disks are updated, but partitions and PVs do not.

After the swact, the conductor will start running on controller-1 with
the new version and will be able to check for `lvm_pv_uuid` before updating the database.

This is the change that prevents the conductor from creating null
entries in the database.

Test Plan:
  PASS: DX - Test USM upgrade procedure changing database state to
  reflect device_node change forcing the error condition
  PASS: DX - Bootstrap and provisioning validation

Closes-Bug: 2093736

Change-Id: I277d674acd79a18f73c69c90fde3caa44f118319
Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>