Enable all sources firewall rule for IGMP.
It was detected that the firewall was allowing IGMP messages with
source IPv4 address with the platform network or 0.0.0.0/32. But
IGMPv3 membership queries do not have a mandatory source address and
its source address may not be part of the platform interface's
network. This prevents the multicast reception for the opened sockets.

This change removes the source address filtering for IGMP.

Test Plan:
=========
[PASS] with the updated code execute node lock/unlock and check the
       new rule by sending IGMP queries in the management and
       cluster-host interfaces with a source address from
       a network not present on the address pools.
[PASS] trigger a runtime execution by creating the file
       /etc/platform/.platform_firewall_config_required, this file
       is detected by sysinv-agent that requests to sysinv-conductor.
       Then test if the membership query is replied

Closes-Bug: 2093329

Change-Id: I762b96dc9d90ba3f02dfd4de8df4c4b70de11e83
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>