Prevents kube-cert-rotation script restart sysinv
The kube-cert-rotation.sh script is executed inside a cron-job
that runs once a day. If this happens during a kubernetes upgrade,
it will cause the upgrade to fail because sysinv is restarted
there.

This adds a check to see whether an upgrade is in progress. If so,
a new attempt is made 1 hour later.

Test Plan:
PASS: Run the kube-cert-rotation.sh script without upgrade in
      progress. The script runs in its entirety.
PASS: Run the kube-cert-rotation.sh script with upgrade in
      progress. The script waits 1 minute and checks the k8s
      upgrade status again. If completed, it continues,
      otherwise the execution fails.

Closes-bug: 2095027

Change-Id: Iff2d25f05c2069f4de2bac65425e1e27094b8c91
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>