Add to agent the same version and usm state check present in conductor
Only allow the Agent to send storage related updates
(disk/part/pvs/lvs/fs) if the state is DEPLOYED and the software
version in the database matches the actual version that can be read
in tsc (/etc/build.info).

Before this change the agent could send information to the conductor
while that conductor is running a different software version, which
could cause the conductor to ignore the message sent. To prevent the
agent from sending storage related updates unnecessarily and doing
more work than it needs to, this improvement is needed.

Since the Agent does not access the database, request the USM state
to the conductor using RPC API.

The conductor caches a variable to track if an upgrade is in progress,
which can be queried by the agents, reducing overhead.

Test Plan:
  - PASS: Perform an upgrade altering the DB to reflect the desired
          states for USM and sw_version.
  - PASS: Perform a complete upgrade and verify that no alarms are
          generated.
  - PASS: Bootstrap and initial provisioning.
  - PASS: Swact controllers.
  - PASS: Lock/Unlock hosts one at a time.
  - PASS: Reboot hosts one at a time.

Test plan executed on a DX and a STORAGE 2+2+1 deployment.

Closes-Bug: 2095440

Change-Id: Ia47876a79e31155cb3727370e0196a8b57db47c7
Signed-off-by: Heitor Farhat <heitor.farhat@windriver.com>