Terminate lingering kubeadm and puppet process during K8S abort
During a Kubernetes orchestration upgrade, if the kubeadm
control plane upgrade process is interrupted, it can cause
the control plane upgrade to fail due to a timeout. If a
Kubernetes abort is triggered after this interruption, the
system will become stuck in the "upgrade-aborting" state.

This change ensures that both the kubeadm and Puppet processes
are properly terminated when a Kubernetes abort occurs.
By doing so, it allows the system to successfully complete
the abort process.

This updates the runtime configuration report status for
update control plane to failed.

Test Plan:
PASS: Perform a manual and orchestrated Kubernetes upgrade
from version 1.24.4 to 1.25.3.
PASS: Execute the Kubernetes orchestration upgrade, and verify
by manually stopping the kubeadm and its parent process during the
control plane upgrade. Ensure that Kubernetes aborts successfully
after the timeout is reached.
PASS: Perform a manual Kubernetes upgrade and verify by aborting
the upgrade at each step, ensuring that it is aborted successfully
at each stage.

Closes-Bug: 2095599

Change-Id: Ia215998a70eb60eee44b2aadbb1ad726d136648b
Signed-off-by: Boovan Rajendran <boovan.rajendran@windriver.com>