Add new clock pin direction parameter for synce4l
A new 'external_source_direction' meta parameter was added to synce4l's
interface parameters' list. It enables the user to define the direction
(function) of the clock pin (input or output).

It will be used for generating the 'external_enable_cmd' and
'external_disable_cmd' values if they are not set manually by the user.

This parameter is optional. When not specified, a default value will
be chosen depending on the clock pin.

Test Plan:
PASS: check the commands generation for each clock pin when
      'external_source_direction' is not set
PASS: check the commands generation for each clock pin when
      'external_source_direction' is set to different values, including
      bad ones
PASS: add 'external_enable_cmd'/'external_disable_cmd' manually with and
      without 'external_source_direction' set

Story: 2010540
Task: 51599

Change-Id: Ibb193f792becfbbafc2fc08f1bdac22d0651c20c
Signed-off-by: Caio Bruchert <caio.bruchert@windriver.com>