Add platform-integ-apps to critical app list
Adding platform-integ-app to the list of critical apps is a
strategy to make better use of time and ensure that while
platform-integ-app is updating, cert-manager's leader election
takes place in parallel allowing it to be ready to issue
certificates. With this, we have improved the waiting time for
the certificate to be ready and in the execution of script itself.

The tests indicated a significant improvement:

 - Without this fix, script 21 took 5m10s to execute completely
   and waited for the certificate to be ready for 53s.

 - With this improvement, script 21 took 4m34s to execute
   completely and waited for the certificate to be ready for only
   20s.

Test Plan:
PASS: build-pkgs && build-image
PASS: upgrade from stx-9 to master
PASS: upgrade performed with the portieris app  which is dependent
      on cert-manager and was successfully updated.

Story: 2011242
Task: 51607

Signed-off-by: David Bastos <david.barbosabastos@windriver.com>
Change-Id: I66f197788062115f24718b6b86f7488e8fe0d8ba