Add super-admin.conf to K8s Root CA update
Certificate super-admin.conf is a new certificate introduced recently in
k8s. This commit adds it to the Kube Root CA update orchestration.

A short extract from [1] explaining the motivation for having
super-admin.conf in addition to admin.conf:

[1] https://github.com/kubernetes/kubernetes/blob/master/CHANGELOG/CHANGELOG-1.29.md

"kubeadm: a separate "super-admin.conf" file is now deployed. The User
in admin.conf is now bound to a new RBAC Group kubeadm:cluster-admins
that has cluster-admin ClusterRole access. The User in super-admin.conf
is now bound to the system:masters built-in super-powers / break-glass
Group that can bypass RBAC. Before this change, the default admin.conf
was bound to system:masters Group, which was undesired."

This change is doing 3 things in summary:

1 - Add super-admin.conf to the orchestration.

2 - Change org names accordingly to reflect K8s default certificates

Besides of this decision on part of k8s, some of the OrgNames were also
changed. That's the case for admin.conf and apiserver-kubelet-client.crt
which had org as system:masters before and now have it as
kubeadm:cluster-admins. Only super-admin.conf now has that
system:masters org. This change updates the Kube Root CA procedure to
use the new org names accordingly.

3 - Refactoring to remove duplication.

This change also refactors method _build_k8s_controller_certificates a
bit. It had about 50 lines of duplication for each certificate. The
refactoring I did here is not radical and only extracts a function for
the cert template and k8s api call. Each certificate is now passed to
this function as an object.

Test case: (Tested on SX and DX)

PASS: Run Kube Root CA update and verify that every certificate
      signed by K8s Root CA is updated in show-certs.sh. For DX,
      do this check in controller-1 too. Verify k8s is operation.
      Run host-lock/unlock and verify k8s and certificates are
      still the same.

Closes-Bug: 2097100
Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/940338

Change-Id: Iaf8218fcf0ab929788700c71b9d622063fb2124a
Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>