Prevents kube-cert-rotation.sh running during platform update
The kube-cert-rotation.sh script is executed inside a cron-job
that runs once a day.

During the platform upgrade, the kubeapi-server may become
unavailable at some stage. If the kube-cert-rotation script is
executed at this time, it will fail and raise the alarm
"Kubernetes certificates renewal failed", which may cause the
upgrade process to fail.

To prevent this, a check is made at the beginning of the code to
see if an upgrade is in progress. If so, a new attempt will be
made in 2 hours. If it fails on the second attempt, the script
will be stopped.

Not executing the script for 1 day is not a problem, since new
attempts are made every day.

Test Plan:
PASS: Run the kube-cert-rotation.sh script without upgrade in
      progress. The script runs in its entirety.
PASS: Run the kube-cert-rotation.sh script with upgrade in
      progress. The script waits 2 hours  and checks the platform
      upgrade status again. If completed, it continues,
      otherwise the execution fails.

Closes-bug: 2095027

Change-Id: Ieedb0b1eb13533c54871001006108d2806eb6c56
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>