Allow null values when patching osdid for rook-ceph
Usually null values are not allowed when patching
an osdid for any backend type.

Now we allow null values for when patching an
osdid with a replace operation while using rook-ceph
backend type.

This is to allow us to use the null value as a "reset"
operation performed in the ceph-osd-sync custom job.

Test Plan:
  PASS: Fresh install
  PASS: Change OSD journal_size and reapply
  PASS: Add OSDs and reapply
  PASS: Remove OSDs and reapply
  PASS: Force altering device_node on OSD deployment
           and reapply

The test plan was executed on a two-node system.

Partial-Bug: 2097133

Change-Id: I09cb99251dc006847b45c018330a3d459d0c551d
Signed-off-by: Heitor Farhat <heitor.farhat@windriver.com>