Reduce manifests running after upgrade completes
The method _controller_config_active_apply is executed
periodically to run pending config manifests if flags are set. These
flags are being set incorrectly during upgrades, leading to
unnecessary reconfiguration when the upgrade completes, taking extra
3~5 minutes to clear out.

This review fixes a condition to remove the flag
'.admin_endpoint_config_required', that was not updated to match
the usm upgrade, and removes the flag '.https_config_required' in
script #102. The flag is being carried from the old config folder,
and the reconfiguration is not required if the from side of the
upgrade already has https enabled. With these changes, the extra
manifests will not be triggered when the from side has https
enabled already.

In the abscense of the flags, a reconfiguration needs to be called
directly to add new services (e.g. dcagent). This will be done by
a new activation script, that will run at the end of the upgrade
activation and perform the reconfiguration using the faster
implementation in sysinv.

Test Plan:
PASS: Bootstrap AIO-SX.

PASS: Bootstrap DC Systemcontroller + SX subcloud.

PASS: Upgrade DC SystemController + SX w/ https enabled.
      Observe that the method _controller_config_active_apply
      doesn't trigger pending updates at the end of the upgrade.

PASS: Upgrade DC SystemController + SX w/ https disabled.
      (Behavior is the same as before)

Closes-bug: 2095442

Change-Id: Ia6bc8b8213861486a0ab6209d16c1c3547e03562
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>