Prevent script from stopping due to command failure
The kube-cert-rotation.sh script is executed when services are down,
so the "system kube-upgrade-show" and "software deploy show"
commands may fail at this point.

This change ensures that the script will not stop executing if
either of these commands fails.

Test Plan:
PASS: Run the kube-cert-rotation.sh script without platform upgrade
      in progress or also without k8s upgrade in progress. The
      script runs in its entirety.
PASS: Block sysinv port 6385 and run the script. The
      "system kube-upgrade-show" command fails, but is unable to
      stop the script from running.

Closes-bug: 2095027

Change-Id: I12f1d98f78faf230ab82cd7ab896825eba8a71e6
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>