Bond if-class restriction and VF sharing
When creating a bond interface with no class, the interface
is not populated by sysinv in the hieradata files for each
host used by puppet. Therefore, restrict interfaces of type
'ae' to only have if-class 'platform' or 'data'.

Also, one is able to create a VF interface using a physical
function and then creating a bond interface using that physical
function as one of the slaves. However, this cannot be done the
other way around. Therefore, allow the creation of a VF that uses
a slave of an existing bond interface.

Testing:
- Build successful
- DX deployment successful
- Ensure an interface of type 'ae' is only created with if-class:
  -> platform
  -> data
- Ensure an interface of type 'vf' is able to use a slave
  from an existing bond interface
- Successfully run the new test cases related to these changes

Closes-Bug: 2098881

Change-Id: I315f4fc7a457d6462828bcb9821b9ff204ed65db
Signed-off-by: Mohammad Issa <mohammad.issa@windriver.com>