Allow subnet overlap between cluster-pod and cluster-service subnets.
The change https://review.opendev.org/c/starlingx/config/+/918961,
related to story 2011027 (IPv4/IPv6 Dual Stack Networking),
refactored the validation between address pool ranges. It started to
block the configuration of overlap between cluster-pod and
cluster-service.

This case is an exception that needs to be allowed, and this change
restores that.

Test Plan
---------
[PASS] unit tests
[PASS] bootstrap system with network overlap
[PASS] upgrade from a system that contained overlap

Closes-Bug: 2099679

Change-Id: Ife44678433e45d3d4ef8f9ac64aaa2c1c9e2051b
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>