Ensure configuration is applied before updating control plane status
During a kubernetes control plane upgrade using the orchestration
method,the status is being updated before the configuration is
applied. This causes the kubelet upgrade to run before the 250.001
alarm is cleared, leading to a failure.

This change ensures that the configuration is successfully applied
before the control plane upgrade status is updated, preventing the
premature execution of the kubelet upgrade process.
In this change we check for the alarm status before updating the
control plane upgrade status in a separate thread.

Test Plan:
PASS: Perform simplex and duplex k8s upgrade successfully.
PASS: Perform an orchestrated Kubernetes upgrade from version 1.24.4
      to 1.29.2
PASS: Manually send STOP signal to pause the process so that
      puppet manifest timeout and check whether retry code works
      and in retry attempts the upgrade completes.
PASS: Manually decrease the puppet timeout to very low number
      and verify that code retries 2 times and updates failure
      state
PASS: Perform a manual Kubernetes upgrade, make the control
      plane upgrade to fail after retries, then manually abort
      and verify it succeeds.
PASS: Perform an orchestrated Kubernetes upgrade, make the control
      plane upgrade fail after retries, and verify that nfv
      triggers and succeeds in aborting the upgrade.

Closes-Bug: 2099665

Change-Id: Ibf740ceab74bcaa80fbbdd1b87e830d0c5fede1a
Signed-off-by: Boovan Rajendran <boovan.rajendran@windriver.com>