Fix auto managed apps being stuck in uploading state.
Some apps were getting stuck in an uploading state if unhandled
exceptions were raised during the auto_upload_managed_apps function.

This commit updates the app status to upload-failed if an exception is
raised during the auto_upload_managed_apps method, and also updates
the condition in the audit method to allow the system to call the
auto_upload function to the failed apps again in the next iteration.

Test plan:
PASS: SX fresh install
PASS: force the system raises an exception during the
      auto_upload_managed_apps execution.
PASS: check if auto managed apps transition from uploading to
      upload-failed status
PASS: check if in the next audit iteration the managed apps were
      installed correctly.

Closes-bug: 2099926

Change-Id: Ic10482f98c38229f205e0cba0e2947a86291455a
Signed-off-by: edias <edson.dias@windriver.com>