Fix reading termination condition for large message
This commit fixes the termination conditions of the loop for reading
message from socket during ipsec-server and ipsec-client communication.
The current conditions will break the loop prematurely for message
bigger than 8192, with the rest of the message lost. With this fix,
the max length of the message will be 10 x 8192 bytes.

Test Plan:
PASS: On a DX system, running ipsec-server and ipsec client in debug
      mode, simulate IPsec initial-auth operation with a message bigger
      than 8192. Verify that the ipsec-client read the whole message,
      and the IPsec initial auth operation competes successfully and
      controller-0 is correctly configured.
PASS: Boostrap a DX system with system-local-ca that has multiple
      layers of ICAs (so that the Initial Auth Response message exceeds
      8192 bytes), verify boostrap succeeds, IPsec certificates are
      generated, and IPsec SAs are established.
PASS: Deploy the systemcontroller for a DC system, after deployment run
      platform certificate update playbook to update system-local-ca
      with multiple layers of ICAs (so that the Initial Auth Response
      message exceeds 8192 bytes). Then add a DX subcloud to the DC,
      verify the subcloud is deployed successfully, and IPsec SAs are
      established in the subcloud.

Closes-Bug: 2100550

Change-Id: I167ff6b2eef0364f50c59e39b73873f7fc177762
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>
Signed-off-by: Andy Ning <andy.ning@windriver.com>