Prepare data with rate limit params for Puppet
In this commit, prepare the data of interfaces details with rate
limiting parameters to pass this data to puppet repo. This code
will be used in the fresh install and it supports both IPv4 and
IPv6.

Rate-limiting is only applicable for external platform networks
and their interfaces.
The following validations are applied on Interfaces:
1. Interface Class is Platform.
2. Interface type is ethernet, vlan or ae.
3. Internal network types like pxeboot, cluster-host are not
   applicable.
4. Mgmt network is applicable for rate-limiting only in case of
   Distributed Cloud setup.
5. External Networks like oam, admin are applicable.

TEST PLANS:

Tested the following cases in
1. AIO-SX
2. Standard
3. DC system Controller and Subcloud machines.
- for controller and worker personalities.

PASSED: Verified after fresh install. There are no default values for
        max_tx_rate and max_rx_rate. tmp/puppet/hieradata/host.yaml
        file is getting updated as expected.

PASSED: Added max_tx_rate and max_rx_rate values in db manually and
        verified these values are properly updated in
        tmp/puppet/hieradata/host.yaml file for different ip versions
        after host unlock. Verified with different combination of
        values including 0. Swact behavior is as per existing design
        (Active controller interface configurations are not reflected
         in the new active controller after the swact operations.)
PASSED: Verify rate-limiting on mgmt interface only allowed in DC
        System configurations.
PASSED: Any rate-limiting on Internal networks is not allowed.

Story: 2011350
Task: 51742
Depends-On: https://review.opendev.org/c/starlingx/config/+/943222

Change-Id: I7b66a94a2a410f19ebba44237d206e2f25720b5b
Signed-off-by: Md Irshad Sheikh <mdirshad.sheikh@windriver.com>